chore: removed original unused version of ConsoleApp
