+++ PV260_Project/PL/ConsoleApps/ConsoleApp.cs
@@ -1,70 +1,71 @@
-﻿using System.Globalization;
+using System.Globalization;
 using BL.DataLoading;
 using BL.DiffComputing;
+using BL.Logging;
 using BL.Writers;
 using CommandLine;
 
 namespace PL.ConsoleApps
 {
-    // > first version of ConsoleApp that doesn't do any 'extensive' logic/parsing in constructor,
-    //   which means we can't initialize _dataLoader / _resultWriter (and only do it 'late' after parsing args)
-    // > improved (?) version, ConsoleAppImproved, does some parsing in constructor so that we can properly
-    //   set _dataLoader / _resultWriter
     public class ConsoleApp : IApp
     {
-        private IDataLoader _dataLoader;
-        private IResultWriter _resultWriter;
+        private readonly IDataLoader? _dataLoader;
+        private readonly IResultWriter? _resultWriter;
         private readonly IDiffComputer _diffComputer = new DiffComputer();
-
-        public void Run()
+        private readonly ILogger _logger = new ConsoleLogger();
+        private readonly Options? _options;
+        
+        public ConsoleApp()
         {
             var args = Environment.GetCommandLineArgs();
-            Parser.Default.ParseArguments<Options>(args)
-                .WithParsed(Execute)
-                .WithNotParsed(HandleParseError);
-        }
-
-        private void Execute(Options o)
-        {
-            if (!AreOptionsValid(o))
+            _options = Parser.Default.ParseArguments<Options>(args).MapResult(ValidateOptions, HandleParseError);
+            if (_options == null)
             {
                 return;
             }
 
-            // can't initialize it earlier
-            _dataLoader = new CsvFileLoader(o.Delim);
-            _resultWriter = (o.OutputFile == null)
-                ? new ConsoleResultWriter(new CultureInfo("en-US"), DateTime.Now, o.Delim)
-                : new FileResultWriter(o.OutputFile, new CultureInfo("en-US"), DateTime.Now, o.Delim);
-
+            _dataLoader = new CsvFileLoader(_options.Delim);
+            _resultWriter = (_options.OutputFile == null)
+                ? new ConsoleResultWriter(new CultureInfo("en-US"), DateTime.Now, _options.Delim)
+                : new FileResultWriter(_options.OutputFile, new CultureInfo("en-US"), DateTime.Now, _options.Delim);
+        }
+        
+        public void Run()
+        {
+            if (_options == null || _dataLoader == null || _resultWriter == null)
+            {
+                return;
+            }
+            
             try
             {
-                var fstParsedFile = _dataLoader.LoadCsvFile(o.InputFilesList[0]);
-                var sndParsedFile = _dataLoader.LoadCsvFile(o.InputFilesList[1]);
+                var fstParsedFile = _dataLoader.LoadCsvFile(_options.InputFilesList[0]);
+                var sndParsedFile = _dataLoader.LoadCsvFile(_options.InputFilesList[1]);
                 var changes = _diffComputer.ComputeDiff(fstParsedFile, sndParsedFile);
                 _resultWriter.Print(changes);
             }
             // TODO: should use 'our' generic exception (that covers errors for reading from disk, loading/fetching from web etc.) ?
             catch (Exception e)
             {
-                Console.WriteLine(e);
+                _logger.Log(e.Message);
                 throw;
             }
         }
-
+        
+        private Options? ValidateOptions(Options o)
+        {
+            return AreOptionsValid(o) ? o : null;
+        }
+        
         private bool AreOptionsValid(Options o)
         {
-            // delimiter can be any string?
-            // outputFile can be any string?
             var inputFiles = o.InputFilesList;
-
-            // Count = 2 is assured by Options.InputFiles (otherwise parse error), therefore useless 'if' - should remove ?
             if (inputFiles.Count != 2)
             {
                 Console.WriteLine("You need to specify exactly TWO input CSV files.");
                 return false;
             }
-
+            
             var invalidPath = false;
             foreach (var inputFile in inputFiles)
             {
@@ -82,11 +83,19 @@ namespace PL.ConsoleApps
             }
 
             return true;
-        }
-
-        private void HandleParseError(IEnumerable<Error> errors)
+        } 
+        
+        private Options? HandleParseError(IEnumerable<Error> errs)
         {
-            Console.WriteLine($"errors {errors.Count()}");
+            var errors = errs.ToList();
+            Console.WriteLine($"errors: {errors.Count}");
+            _logger.Log($"CommandLineParser - {errors.Count} errors:");
+            foreach (var e in errors)
+            {
+                _logger.Log($" - Error: {e.ToString() ?? "CommandLine." + e.Tag}");
+            }
+            
+            return null;
         }
     }
 }
